+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## 1.0.0
+- Detect when the screen is locked
+- Detect when the screen is unlocked
+- Supports both event streams and manual status checks
+- Works on Android and iOS